update music app

+++ webapps/webapp_music/api/remove.php
@@ -1,5 +1,5 @@
 <?php
-//echo "<pre>";
+//echo "<spanre>";
 //print_r($_REQUEST);
 //echo "</pre>";
 //exit;
@@ -42,23 +42,24 @@ for( $n1=0; $n1 < count($file); $n1++ ){
 
 	if( is_dir($filename) ){
 		if( rmdir( $filename ) ){
-			$logMsg["message"] .= "<p class='msg-success'><b>rmdir empty folder</b> $filename</p>";
+			$logMsg["message"] .= "<span class='msg-success'><b>rmdir empty folder</b> $filename</span><br>";
 		} else {
 			$res=false;
 			$res=RemoveTree($filename);
 			if ($res){
-				$logMsg["message"] .= "<p class='msg-success'><b>remove tree</b> $filename</p>";
+				$logMsg["message"] .= "<span class='msg-success'><b>remove tree</b> $filename</span><br>";
 			} else {
-				$logMsg["message"] .= "<p class='msg-warning'><b>cannot rmdir</b> $filename</p>";
+				$perms = substr(sprintf('%o', fileperms( $filename ) ), -4);
+$logMsg["message"] .= "<span class='msg-warning'><b>cannot rmdir</b> $filename, access rights: $perms</span><br>";
 			}
 		}
 	}
 
 	if ( is_file($filename) ){
 		if (unlink ($filename)){
-			$logMsg["message"] .= "<p class='msg-success'><b>delete</b> $filename</p>";
+			$logMsg["message"] .= "<span class='msg-success'><b>delete</b> $filename</span><br>";
 		} else {
-			$logMsg["message"] .= "<p class='msg-error'><b>cannot delete</b> $filename</p>";
+			$logMsg["message"] .= "<span class='msg-error'><b>cannot delete</b> $filename</span><br>";
 		}
 	}
 
@@ -74,29 +75,37 @@ function RemoveTree($dir)
 { 
 	global $logMsg;
 
-	$handle = opendir($dir) or die("Cannot open directory $dir"); 
-	while (false !== ($file = readdir($handle))) 	{ 
-		if ($file != "." && $file != "..") { 
-			if( is_file($dir."/".$file) ) { 
-				if( unlink($dir."/".$file) ) {
-					$logMsg["message"] .= "<p class='msg-success'><b>delete</b> $file</p>";
+	//$handle = opendir($dir) or die("Cannot open directory $dir"); 
+	if ($handle = opendir($dir) ){
+
+		while (false !== ($file = readdir($handle))) 	{ 
+			if ($file != "." && $file != "..") { 
+				if( is_file($dir."/".$file) ) { 
+					if( unlink($dir."/".$file) ) {
+						$logMsg["message"] .= "<span class='msg-success'><b>delete</b> $file</span><br>";
+					} 
 				} 
-			} 
-			if(is_dir($dir."/".$file)) { 
-				RemoveTree($dir."/".$file);
-				if( rmdir($dir."/".$file) ) {
-					$logMsg["message"] .= "<p class='msg-success'><b>rmdir</b> $dir."/".$file</p>";
+				if(is_dir($dir."/".$file)) { 
+					RemoveTree($dir."/".$file);
+					if( rmdir($dir."/".$file) ) {
+						$logMsg["message"] .= "<span class='msg-success'><b>rmdir</b> $dir."/".$file</span><br>";
+					} 
 				} 
+				
 			} 
-			
+		}//end while
+		closedir($handle); 
+		
+		if(rmdir($dir)){
+			$logMsg["message"] .= "<span class='msg-success'><b>rmdir</b> $dir</span><br>";
+			return true;
 		} 
-	} 
-	closedir($handle); 
+		
+	} else {
+		$perms = substr(sprintf('%o', fileperms( $dir ) ), -4);
+		$logMsg["message"] .= "<span class='msg-error'>Cannot open file system object $dir, access rights: $perms</span><br>";
+	}
 	
-	if(rmdir($dir)){
-		$logMsg["message"] .= "<p class='msg-success'><b>rmdir</b> $dir</p>";
-		return true;
-	} 
 }//end
 
 ?>
